AJAX to submit the Google Form
Following the ideas in https://gohkhoonhiang.github.io/2016/05/simple-contact-form-on-jekyll/
to submit tfe form data to Google Forms directly, without an iframe embed

When the form submit's by default, Google Forms redirects the user to a "Signup Completed" page.
To prevent this, used JavaScript to override the submit button's action.
Had to move the inialization of jQuery to the top of default page, otherwise can't use document.ready

NOTE: The form is linked to my personal Google account
Once this is live, the url variable and "entry.XXXXXXXX" fields need to be changed